add libraries
